Added protocols folder

+++ protocols/manifests/init.pp
@@ -1,29 +1,29 @@
-class protocols ($protocols) {
+class protocols { 
+require java
+$token = "TOKEN"
 
-# one-off defined resource type, in
-# /etc/puppetlabs/code/environments/production/modules/puppet/manifests/binary/symlink.pp
+#FACTER_protocols="Protocol1Server","Protocol2Server","Protocol3Server" puppet agent --test
+
 define puppet::binary::symlink ($protocol = $title) {
-  $token = gentoken("ex1${protocol}")
-  file {"${protocol}":
-    path => "/tmp/${protocol}.java",
-    content => template("protocols/${protocol}.java.erb"),
-    #target => "/opt/puppetlabs/bin/${binary}",
-    owner => 'root',
-    group => 'root',
-    mode => '0700',
-    before => Exec["compile_server${protocol}"]
-  }
-  exec { "compile_server${protocol}":
-	command => "javac ${protocol}.java &&
-	mv ${protocol}.class /root/${protocol}.class &&
-	rm ${protocol}.java",
-        path => '/usr/bin/:/bin',
-	cwd => '/tmp',
-  }
+if("${protocol}" != '') {
+	  file {"${protocol}":
+	    path => "/tmp/${protocol}.java",
+	    content => template("protocols/${protocol}.java.erb"),
+	    owner => 'root',
+	    group => 'root',
+	    mode => '0700',
+	    before => Exec["compile_server${protocol}"]
+	  }
+	  exec { "compile_server${protocol}":
+		command => "javac ${protocol}.java &&
+		mv ${protocol}.class /root/${protocol}.class &&
+		rm ${protocol}.java",
+		path => '/usr/bin/:/bin',
+		cwd => '/tmp',
+	  }
+	}
+}
+$protocolsArray = split($::protocols, '[,]')
+puppet::binary::symlink { $protocolsArray: }
 }
 
-# using defined type for iteration, somewhere else in your manifests
-#$protocols = ["Protocol1Server", "Protocol2Server"]
-
-puppet::binary::symlink { $protocols: }
-}
